Add normlize_longitude static method to Domain
normlize_longitude is used to shift longitude to a faraway reference
value, so that the domain is as monotonic as it can be. It also changes
North Pole longitude if it is an extrema, so that it does not affect
the actual range of longitude of the domain.

The new method is used in both create_subdomains and create_mask_domain